Added the ability to run customizable init scripts from /etc/dainit
The scripts have a simple markdown-like format. They start with
 # ServiceName
to identify the script, and then contain freetext. If any line
starts with "Needs:" or "Provides:", it will interpret it as a
(anded) list of service types that the service either depends on,
or provides. If the line starts with "Startup:" the rest of the line
will be interpreted as a command to run to startup the service (after
all Needs are met.) "Shutdown:" is also parsed, but not yet implemented.